project-3/requirements/base.txt
fastparquet==2023.8.0
fire==0.5.0
google-auth==2.23.3
gspread==5.11.3
gspread-dataframe==3.3.1
pydantic==2.4.2
pandas==2.0.3
pytest==7.4.2
python-decouple==3.8
black==23.11.0
pylint==3.0.2
python-dotenv==1.0.0